chore: added @ alias and updated all components

+++ src/service/user-service.ts
@@ -1,14 +1,14 @@
-import { AppError } from "../AppError";
 import httpStatus from "http-status";
 import bycrypt from "bcrypt";
-import { IUser } from "../types/user.types";
-import { UserModel } from "../model/user-model";
-import { AuthUser } from "../types/auth-user.types";
 import { Request } from "express";
 import fs from "fs";
-import { saveLocalUpload } from "../utils/image-upload";
 import path from "path";
-import { logger } from "../logger";
+import { IUser } from "@/types/user.types";
+import { AppError } from "@/AppError";
+import { UserModel } from "@/model/user-model";
+import { AuthUser } from "@/types/auth-user.types";
+import { saveLocalUpload } from "@/utils/image-upload";
+import { logger } from "@/logger";
 
 export const createUserService = async (requestUser: IUser) => {
   if (!requestUser.name || !requestUser.email) {
@@ -44,11 +44,13 @@ export const saveUsernameService = async (
   try {
     const photoUrl = saveLocalUpload(req.file);
     existingUser.username = username;
-    existingUser.photo = photoUrl ?? undefined;
+    if (photoUrl) {
+      existingUser.photo = photoUrl;
+    }
 
     await existingUser?.save();
 
-    if (oldUserPhoto) {
+    if (photoUrl && oldUserPhoto) {
       const oldPath = path.join(process.cwd(), oldUserPhoto);
       fs.unlink(oldPath, (err) => {
         if (err) logger.error("Failed to delete old photo:", err);
